% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{AIC.simplexregression}
\alias{AIC.simplexregression}
\title{Akaike Information Criterion}
\usage{
\method{AIC}{simplexregression}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}

\item{k}{Penalty parameter (default: 2)}
}
\value{
Numeric AIC value
}
\description{
Computes the AIC for a simplex regression model.
}
