% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{BIC.simplexregression}
\alias{BIC.simplexregression}
\title{Bayesian Information Criterion}
\usage{
\method{BIC}{simplexregression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Numeric BIC value
}
\description{
Computes the BIC for a simplex regression model.
}
