% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{coef.simplexregression}
\alias{coef.simplexregression}
\title{Extract Coefficients}
\usage{
\method{coef}{simplexregression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Named numeric vector with all coefficients (mean, dispersion, lambda)
or (mean, dispersion)
}
\description{
Extracts all coefficients from a simplex regression model.
}
