% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{df.residual.simplexregression}
\alias{df.residual.simplexregression}
\title{Extract Residual Degrees of Freedom}
\usage{
\method{df.residual}{simplexregression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Integer residual degrees of freedom
}
\description{
Extracts the residual degrees of freedom.
}
