% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{fitted.simplexregression}
\alias{fitted.simplexregression}
\title{Extract Fitted Values}
\usage{
\method{fitted}{simplexregression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Numeric vector of fitted mean values
}
\description{
Extracts fitted values from a simplex regression model.
}
