% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{formula.simplexregression}
\alias{formula.simplexregression}
\title{Extract Model Formula}
\usage{
\method{formula}{simplexregression}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Formula object
}
\description{
Extracts the model formula from a simplex regression model.
}
