% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_influence.R
\name{gleverage_simplexreg}
\alias{gleverage_simplexreg}
\title{Generalized leverage for Simplex Regression}
\usage{
gleverage_simplexreg(model)
}
\arguments{
\item{model}{An object of class \code{"simplexregression"}}
}
\value{
Numeric vector of leverage values
}
\description{
Computes generalized leverage measures for simplex regression
model with Parametric or Fixed Link.
}
