% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_influence.R
\name{hatvalues.simplexregression}
\alias{hatvalues.simplexregression}
\title{Hat Values for Simplex Regression}
\usage{
\method{hatvalues}{simplexregression}(model, ...)
}
\arguments{
\item{model}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Numeric vector of hat values
}
\description{
Computes hat (leverage) values for the mean submodel.
}
