% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{logLik.simplexregression}
\alias{logLik.simplexregression}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{simplexregression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Object of class \code{"logLik"} with attributes df and nobs
}
\description{
Extracts the log-likelihood from a fitted model.
}
