% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{model.matrix.simplexregression}
\alias{model.matrix.simplexregression}
\title{Model Matrix Extraction}
\usage{
\method{model.matrix}{simplexregression}(object, type = c("mean", "dispersion"), ...)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{type}{Character string indicating which design matrix to extract:
\code{"mean"} (default) or \code{"dispersion"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Model matrix (numeric matrix)
}
\description{
Extracts the design matrix used in the mean or dispersion submodel.
}
