% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{nobs.simplexregression}
\alias{nobs.simplexregression}
\title{Extract Number of Observations}
\usage{
\method{nobs}{simplexregression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Integer number of observations
}
\description{
Extracts the number of observations used in model fitting.
}
