% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{print.simplexregression}
\alias{print.simplexregression}
\title{Print Method for Simplex Regression with Parametric Link}
\usage{
\method{print}{simplexregression}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class \code{"simplexregression"}}

\item{digits}{Number of digits to print}

\item{...}{Additional arguments (currently not used)}
}
\description{
Prints a concise summary of a fitted simplex regression model.
}
