% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_resettest.R
\name{reset_simplexreg}
\alias{reset_simplexreg}
\title{RESET Test for Simplex Regression}
\usage{
reset_simplexreg(model, dispersion = TRUE)
}
\arguments{
\item{model}{An object of class \code{"simplexregression"}}

\item{dispersion}{Logical. If \code{TRUE}, includes the squared linear predictor
in the dispersion submodel as well. Default is \code{TRUE}}
}
\value{
An object of class \code{"htest"} containing:
\itemize{
  \item \code{statistic}: The likelihood ratio test statistic
  \item \code{parameter}: Degrees of freedom
  \item \code{p.value}: The p-value of the test
  \item \code{method}: Description of the test
  \item \code{data.name}: Name of the model object
}
}
\description{
Performs the RESET (Regression Equation Specification Error Test)
to detect functional form misspecification in simplex regression models.
}
\details{
The RESET test augments the original model by adding the squared linear
predictor as an additional covariate. Under the null hypothesis of correct
functional form, this additional term should not be significant.

The test statistic follows a chi-squared distribution with degrees of freedom
equal to the difference in the number of parameters between the augmented
and original models.

If \code{dispersion = TRUE}, the squared linear predictor is added to both
the mean and dispersion submodels. If \code{FALSE}, it is only added to the
mean submodel.
}
\examples{
\dontrun{
# Fit a simplex regression model
model <- simplexreg(y ~ x1 + x2, data = mydata)

# Perform RESET test
reset.simplexregression(model)

# RESET test only for mean submodel
reset.simplexregression(model, dispersion = FALSE)
}

}
\references{
Ramsey, J. B. (1969). Tests for specification errors in classical linear
least-squares regression analysis. \emph{Journal of the Royal Statistical
Society: Series B}, 31(2), 350-371.
}
