% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_loglik.R
\name{simplex_loglik}
\alias{simplex_loglik}
\title{Log-likelihood for the Simplex Regression}
\usage{
simplex_loglik(theta, y, x1, z1, link.mu, link.sigma2)
}
\arguments{
\item{theta}{Parameter vector}

\item{y}{Response vector (0 < y < 1)}

\item{x1}{Design matrix for mean model (with intercept)}

\item{z1}{Design matrix for dispersion model (with intercept)}

\item{link.mu}{Mean link function ("logit", "probit", "cloglog", "loglog",
"cauchit", "plogit1" or "plogit2")}

\item{link.sigma2}{Dispersion link function ("log", "sqrt", or "identity")}
}
\value{
Numeric scalar (log-likelihood)
}
\description{
Computes the log-likelihood function for given parameters in the
  simplex regression models with parametric or fixed mean link functions.
}
\details{
The log-likelihood function for the simplex regression model with
parametric or fixed mean link function is given by
\deqn{
  \ell(\boldsymbol{\theta}) =
  -\frac{1}{2} \sum_{i=1}^{n}
  \left[
    \log(2\pi) +
    \log(\sigma_i^2) +
    3\log(y_i(1 - y_i)) +
    \frac{d(y_i; \mu_i)}{\sigma_i^2}
  \right],
}
where \eqn{d(y_i; \mu_i) = \frac{(y_i - \mu_i)^2}{y_i(1 - y_i)\mu_i^2(1 - \mu_i)^2}} is the
deviance component of the simplex model, \eqn{\mu_i} is obtained from the mean link function,
and \eqn{\sigma_i^2} from the dispersion link function.

The parameter vector \eqn{\boldsymbol{\theta}} has different structures
depending on the type of mean link:
\itemize{
  \item For parametric mean link functions ("plogit1" or "plogit2"):
    \eqn{\boldsymbol{\theta} = (\boldsymbol{\beta}^\top, \boldsymbol{\gamma}^\top, \lambda)^\top},
    with \eqn{p + q + 1} elements.
  \item For fixed mean link functions ("logit", "probit", "cloglog", "loglog", "cauchit"):
    \eqn{\boldsymbol{\theta} = (\boldsymbol{\beta}^\top, \boldsymbol{\gamma}^\top)^\top},
    with \eqn{p + q} elements.
}

Here, \eqn{\boldsymbol{\beta}} and \eqn{\boldsymbol{\gamma}} are regression
coefficients for the mean and dispersion submodels, respectively.
}
