% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_score.R
\name{simplex_score}
\alias{simplex_score}
\title{Score Vector for the Simplex Regression}
\usage{
simplex_score(theta, y, x1, z1, link.mu, link.sigma2)
}
\arguments{
\item{theta}{Parameter vector}

\item{y}{Response vector (0 < y < 1)}

\item{x1}{Design matrix for mean model (with intercept)}

\item{z1}{Design matrix for dispersion model (with intercept)}

\item{link.mu}{Mean link function ("logit", "probit", "cloglog", "loglog",
"cauchit", "plogit1" or "plogit2")}

\item{link.sigma2}{Dispersion link function ("log", "sqrt", or "identity")}
}
\value{
Numeric vector (gradient)
}
\description{
Computes the score vector for given parameters in the
  simplex regression models with parametric or fixed mean link functins.
}
\details{
The parameter vector \eqn{\boldsymbol{\theta}} has different structures
depending on the type of mean link:
\itemize{
  \item For parametric mean link functions ("plogit1" or "plogit2"):
    \eqn{\boldsymbol{\theta} = (\boldsymbol{\beta}^\top, \boldsymbol{\gamma}^\top, \lambda)^\top},
    with \eqn{p + q + 1} elements.
  \item For fixed mean link functions ("logit", "probit", "cloglog", "loglog", "cauchit"):
    \eqn{\boldsymbol{\theta} = (\boldsymbol{\beta}^\top, \boldsymbol{\gamma}^\top)^\top},
    with \eqn{p + q} elements.
}

Here, \eqn{\boldsymbol{\beta}} and \eqn{\boldsymbol{\gamma}} are regression
coefficients for the mean and dispersion submodels, respectively.
}
