% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_fit.R
\name{simplexreg}
\alias{simplexreg}
\title{Simplex Regression with Parametric or Fixed Mean Link}
\usage{
simplexreg(
  formula,
  data,
  diag = 1,
  link.mu = c("plogit1", "plogit2", "logit", "probit", "loglog", "cloglog", "cauchit"),
  link.sigma2 = c("log", "sqrt", "identity")
)
}
\arguments{
\item{formula}{A two-part formula: y ~ mean_predictors | dispersion_predictors}

\item{data}{A data frame containing the variables in the model}

\item{diag}{Diagnostic level (default = 1)}

\item{link.mu}{Mean link function (parametric: "plogit1", "plogit2"; fixed:
"logit", "probit", "loglog", "cloglog", "cauchit")}

\item{link.sigma2}{Dispersion link function ("log", "sqrt", "identity")}
}
\value{
An object of class \code{"simplexregression"} containing:
\itemize{
  \item \code{coefficients}: List with mean, dispersion, and lambda estimates
  \item \code{fitted.values}: Fitted mean values
  \item \code{sigma2.fv}: Fitted dispersion values
  \item \code{mu.lp}: Predicted linear mean
  \item \code{sigma2.lp}: Predicted linear dispersion
  \item \code{optim}: List with initial values and convergence of optim
  \item \code{residuals}: Quantile residuals
  \item \code{vcov}: Variance-covariance matrix
  \item \code{loglik}: Log-likelihood value
  \item \code{AIC}, \code{BIC}, \code{HQ}: Information criteria
  \item \code{R2_N}, \code{R2_FC}: Pseudo R-squared measures

}
}
\description{
Fits a simplex regression model with parametric or fixed mean link
functions.
}
\details{
The model is specified using a two-part formula:
\itemize{
  \item Mean submodel: \eqn{g(\mu_i, \lambda) = x_i'\beta}
  or \eqn{g(\mu_i) = x_i'\beta}
  \item Dispersion submodel: \eqn{h(\sigma^2_i) = z_i'\gamma}
}

The parametric mean link functions include a parameter \eqn{\lambda} that
is estimated along with other model parameters.
}
\examples{
# Simulate data
n <- 100
x1 <- runif(n, 0, 1)
x2 <- runif(n, 0, 1)
mu <- parametric_mean_link_inv(0.8 - 1.2*x1 - 1.5*x2 , 0.25, "plogit2")
sigma2 <- 0.5
summary(mu)
y <- rsimplex_opt(n, mu, sigma2)
data <- data.frame(y = y, x1 = x1, x2 = x2))

# Fit model
fit <- simplexreg(y ~ x1 + x2 | 1, data = data,
                     link.mu = "plogit2", link.sigma2 = "identity")
summary(fit)

}
\references{
Justino, M. E. C. and Cribari-Neto, F. (2025).
Simplex regression with a flexible logit link: Inference and application
to cross-country impunity data.
\emph{Technical report}, Universidade Federal de Pernambuco, Brazil.
}
