% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{terms.simplexregression}
\alias{terms.simplexregression}
\title{Terms Extraction}
\usage{
\method{terms}{simplexregression}(x, type = c("mean", "dispersion"), ...)
}
\arguments{
\item{x}{An object of class \code{"simplexregression"}}

\item{type}{Character string indicating which terms to extract:
\code{"mean"} (default) or \code{"dispersion"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
An object of class \code{"terms"}
}
\description{
Extracts the terms object for the mean or dispersion submodel.
}
