% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{vcov.simplexregression}
\alias{vcov.simplexregression}
\title{Extract Variance-Covariance Matrix}
\usage{
\method{vcov}{simplexregression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"simplexregression"}}

\item{...}{Additional arguments (currently not used)}
}
\value{
Variance-covariance matrix
}
\description{
Extracts the variance-covariance matrix of parameter estimates.
}
